Add routing so boards show for /b/:boardId
Not sure whether to use `/b/:boardId` or `/:boardId`.
The former is better because it carves out a namespace that will never clash
with other routes we may want in the future (such as /my-account or something)
but the later looks nicer.

This currently lets you use whatever board ID in the URL you want as it will be
used in the Firebase ref and the key will be dynamically created upon use.
Not sure if this is good or bad :/

<App> previously had the logic for how a board works but that was fine because
it didn't really have anything else to be responsible for, but now <App> is
responsible for routing and the board logic is in its own component.

The <Board> component should probably have more refactoring as it has a lot of
responsibilities but it'll do as is for now.