Upgrade react-scripts and add a robots.txt
The version of react-scripts we were using was pretty old!
That old version doesn't have the concept of a public folder which meant there
was no easy way of adding a robots.txt

The robots.txt file is important so the homepage gets indexed by google but
individual boards do not - this is mainly for privacy reasons.

Once this feature has been deployed, it should be tested with a test tool:
http://www.google.com/support/webmasters/bin/answer.py?answer=35237