WIP: Move Firebase logic up to Board
This design will make it easier to extract the Firebase logic into yet another
location that can have the responsibility of communicating with the backend.

Eventually, this will make it easier to swap Firebase out for an in house
backend, something like a self-hosted couchDB or a node server with MongoDB and
socketIO.

There is also packages that will connect firebase to a redux store so the app
can use redux as per normal, but more research is required!
https://medium.com/@prescottprue/firebase-with-redux-82d04f8675b9