Make the app bar stick to the top
Following on from the last change, the functionality for adding a column is not
as obvious as it was before, especially if it is hidden at the top of the page
while the user is scrolled down.

I'm not sure if this is a good idea or not but having this in its own commit
makes it easy to revert later.