Prevent the "add card" button from being inset
This one is really subtle, but when you hover over the "add card" button, the
button is smaller that what it looks like and causes the actual button to turn
grey and the padding around it to stay white.

I mainly came across this when I was experimenting with changing the color of
that button and only the middle bit changed.