Make the delete column button less prominent
It is a little weird that we use a trash icon to delete cards and a cross icon
to delete columns, at least this makes the title of the column stand out a bit
more.

Another idea could be to add a kebab menu in the place of this cross button but
I'd imagine that would only be worth it if there were more than one operation to
perform on a column, which I don't think there is yet.