Add merge cards functionality
This includes a refactor to Backend and a change in behaviour for the column so
you can drop a card on the column header instead of anywhere in the column.
This is because it's hard to tell if your dropping it on the card or the
column and so it doesn't work.

Still needs better visual indication but will render a folder icon if the card
has subcards

This will be populated after a drag and drop has occured to turn a card into a
group of cards

:pear: Dean!