Automatically sort cards by votes
According to issue #14, people want to be able to sort cards.

Right now, the sort order is just "order of insertion" but in the happy path,
typical use case, I'd imagine the most useful sort order is by votes.

Before implementing the more advanced option of custom sorting (which will
involve changes to the data model), making the default order sorted by votes
will probably deliver most of the value with much less of the work