Refactor the world! Part 1: Column operations
Refactoring to use a globally available, write only backend.
This should act as a dispatcher of sorts.

To make it globally available, I'm using `react-broadcast` to put the instance
of Backend into context such that:

  <FirebaseProvider> will provide state to the app
  <BackendActions> will provide the instance to complete actions

This may be a bad idea but its better than passing a callback for every
operation in the system (!) and its easier to wire up than redux + react-redux +
redux-thunk (or equivalent)

So far the only part that is working is the operations that a column kicks off
itself:
  - editColumn
  - removeColumn
  - addCard