Fix voting for subcards
This broke as part of a refactor because I forgot to do the second half!
Cards can now take care of telling the backend to vote for themselves.

This is loosely following the "dispatch" model where any component can dispatch
an action as a write-only, CQRS-style command but can not ready anything,
instead that is done via the top-level provider.