Make it easier to write markdown when adding cards
A bigger text field that supports multiline input makes this much easier and the
CSS for the images is to prevent the image from overflowing off the side of the
card and scale down as required.

The edit card functionality is still only a basic browser prompt and should be
replaced with the add card dialogue, then this feature is basically done.