Fix sorting based on subCards
Added the `yarn test` command back, tests helped with the refactor!

The test case explains the problem that was happened before this change,
basically sorting would only take into consideration the votes of a card and not
the sum of the votes + the votes of the subcards.
The UI would still correctly show the total votes but this wasn't used for
sorting until now.