Attempt to log pageviews to GA upon clientside nav
When following a <Link> and letting react-router change pages, the GA code does
not fire again because its already loaded in the page.
This change does fire for every route change and is supposed to send a pageview
to GA.

The problem is even though the function is fired for every route change (as
evident by the console.log lines), looking at the GA realtime dashboard does not
reflect these pages views, only the full page load ones.

So this is currently not working as intended... not sure why yet