Add rich preview meta data through react helmet
The `og` part of the meta data stands for `open graph`. It's a standard created by
facebook for sharing content rich links. Slack has adopted it and is where most
of the time our website is shared. Twitter has its own standard but we probably
don't care about that for now.